Use docker-compose to run spec on Travis
Travis provides elasticsearch 5 by default, which causes errors. If we
want to use 7.4 in this context we'd have to build `before_install`
scripts to install it. If we're going to have to jump through hoops
anyway to use a new elasticsearch, we may as well just use
docker-compose here.

This does make the build run slower (it takes about 6 minutes), but I
think this is acceptable. It also provide parity with the local
development and testing environments used by developers.

This commit also removes the empty test directory (this app uses rspec).
This commit was accidentally rebased into this squashed commit.