Call .to_unsafe_h in category params helper
Controller params are now instances of ActionController::Parameters instead of
HashWithIndifferentAccess. This method just blindly converts the
Parameters to a hash so the category can be stripped out. This is safe
here since this user-supplied input is just used to rende a link on the
page.

There's likely a simpler way to handle this case, but this change allows
us to just upgrade Rails.