Make trial keywords display optional
This adds a new option in the system administration section to indicate
whether trial keywords should be displayed in search results. This
defaults to `true` to maintain backward compatibility.

When this option is set to `false` keywords are still stored and used
for searches, but they are not displayed in results.

Note that because search results are the only place in the application
that these keywords are displayed, if this option is turned off you
can't view a trial's keywords (even in the admin interface). We may
eventually want to add an administrative trial view so keywords (and
all associated trial data) can be viewed by system administrators.