Use Oracle
This replaces Postgres with Oracle for use at the University of
Minnesota.

There's an effort to stand up a Postgres database for this (and other)
apps, but for now we need it to run on Oracle. If and when this moves to
Postgres, this commit should be reverted or removed.