Use select2 tags for Subgroups
This is how the application previoiusly worked. One benefit of using
select2 for this is it handles the parsing of multiple items and special
characters that may be contained in the group names.

The destroy_all here isn't ideal, but feel low risk because this
controller and data is simple, and this is how it's been working.