Move search symbol replacements to Trial.execute_search
It should not be the caller's responsibility to know while symbols need
replacing. This moves the replacement of `/` (which causes an error with
Elasticsearch) to the search method.

Replacement of `"` was removed so terms can be grouped.

Replacement of `head ache` with `headache` has been moved to the
synonyms file.