Always set visible flag based on current values
This changes the ctgov importer to always consider the current
`overall_status` value when determining visibility.

Previously the visible flag was only updated it the parser detected a
change in the overall status. We've found that in production these
can get out of sync; the overall_status becomes something other than
"Recruiting", but the visible flag remains set to true.

The deployment of this change should be accompanied with a one-time
manual update of all trials' visibility flag and a re-index to catch
trials that may no longer be returned from the clinicaltrials.gov
import.