Add all studies API endpoint
Most integrations will want to know all of the NCT numbers that
currently exist in Study Finder. The common pattern is to pull data from
clinicaltrials.gov and then augment those studies with data from other
systems.

This add an index action the the API studies controller. This is a naive
first implementation (some implementations might have too much data for
this to be practical without paging or limiting the fields returned).
The idea is to try this out and change when/if needed.