Remove elasticsearch initializer
This file is not necessary. The elasticsearch gems look for host
configuration in the ELASTICSEARCH_URL environment variable by default
and it defaults to localhost:9200 if nothing is present.

Another problem with this initializer is that it sets log output to
true for all environments. If you won't want Elasaticsearch to log all
transactions in in production for example, you'd currently have override
this initializer somewhere. This change flips this around. By default,
it won't log anything. If you want to enable this during development,
you can opt-in.