Tweak the show page handling of descriptions
This re-arranges some elements on the show page to let the title and
description breathe a bit.

The simple_description helper was replaced with a partial. I'd really
like to push up some of this logic to the model and replace the helpers
with plain old partials—it's super awkward to pass a render method to a
render method. When/if that happens we can pull out the gross
logic from this new partial.