Conditionally render clinicaltrials.gov links
`system_id` isn't always an NCT number. This change will only render the
`See this study on clinicaltrials.gov` link when `system_id` is an NCT
number.

The clinicaltrials.gov link in the "Share via email" email has been
replaced with a StudyFinder link. This will either render a link to the
study show page (if enabled) or a single search result.