Keywords from Oncore annotations not recognized
* "racism" "African American" were not appearing in ElasticSearch results
* "rac" "africa america" and variations excluding the stems of the words would appear
* Determined it was due to "stemming" https://www.elastic.co/guide/en/elasticsearch/reference/current/stemming.html
* Added an override for each individual words
* Validated that "racism" and "african american" now populate intended results

Note: This override most likely is not the best means to an end, as it is additional business logic being forced on top of the Elastic Search logic.
It probably would make more sense to do an entire review of other words (ending in "-an" or "-ism" or others) to verify that the population of Keywords
are not being impacted in other places.

[SF-182]