Temporary mailer disable
Mailer to alert admin to 'bad' system IDs is currently sending huge numbers of emails because the integration is iterating over all studies, including old ones with 'weird' data. Disable the mailer until we update the integration to only look at studies matching the 'visible' criteria for data import, and studies that are currently visible in SF for 'should this study now be hidden?' checks.